pyproject.toml
[project]
name = "quadro"
version = "0.1.0"
description = "Manage your tasks directly from the terminal using markdown"
requires-python = ">=3.12"
dependencies = [
    "click>=8.1.7",
    "rich>=13.7.0",
    "python-frontmatter>=1.1.0",
]

[project.scripts]
quadro = "quadro.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = [
    "ruff>=0.1.0",
    "mypy>=1.7.0",
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "bandit>=1.7.5",
    "pre-commit>=3.5.0",
]

[tool.ruff]
line-length = 100
target-version = "py312"
exclude = [".git", "__pycache__", ".mypy_cache", ".pytest_cache", ".ruff_cache"]

[tool.ruff.lint]
select = ["ALL"]
ignore = ["D", "COM812", "ISC001"]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["S101", "PLR2004", "SLF001"]

[[tool.mypy.overrides]]
module = "tests.*"
disable_error_code = ["misc"]

[tool.ruff.lint.isort]
force-single-line = true
lines-after-imports = 2

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_any_unimported = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_unreachable = true
strict_equality = true
extra_checks = true

[tool.bandit]
targets = ["quadro"]
exclude_dirs = ["tests", ".venv"]
skips = []
severity = "medium"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_functions = "test_*"
addopts = "-v --strict-markers --cov=quadro --cov-report=term-missing --cov-report=html --cov-fail-under=80"
markers = [
    "slow: marks tests as slow",
    "integration: marks tests as integration tests",
]

[tool.coverage.run]
branch = true
source = ["quadro"]
omit = ["tests/*", "**/__init__.py"]

[tool.coverage.report]
precision = 2
show_missing = true
skip_covered = false
